Funguje ale pomale

+++ DataAggregation/Utils.cpp
@@ -7,7 +7,7 @@ using namespace std;
 vector<string> strSplit(string str, string token) {
 	vector<string>result;
 	while (str.size()) {
-		int index = str.find(token);
+		size_t index = str.find(token);
 		if (index != string::npos) {
 			result.push_back(str.substr(0, index));
 			str = str.substr(index + token.size());
